+++ custom_components/aarhuswaste/coordinator.py
