+++ custom_components/aarhuswaste/domain.py
