+++ custom_components/aarhuswaste/manifest.json
@@ -7,5 +7,7 @@
     "config_flow": true,
     "codeowners": [
         "@kvedes"
-    ]
+    ],
+    "icon": "icons/aarhus.png",
+    "logo": "icons/aarhus.png"
 }
