+++ custom_components/aarhuswaste/model.py
